@@ -2,9 +2,7 @@ import { Restaurant } from "./restaurant/restaurant.model";
 
 import { HttpClient, HttpParams } from "@angular/common/http";
 import { Injectable } from "@angular/core";
-import { Observable } from "rxjs/Observable";
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import { Observable } from "rxjs";
 
 import { MEAT_API } from "../app.api";
 
